fix: 将cpp里的money调整为string防止失精，将kt的billInfo里的money调整为Double

+++ app/src/main/java/net/ankio/auto/utils/server/model/BillInfo.kt
@@ -36,7 +36,7 @@ class BillInfo {
     /**
      * 金额 大于0
      */
-    var money: Int = 0
+    var money: Double = 0.0
 
     /**
      * 手续费
